clean: move application DSL
