+++ src/main/scala/com/nulabinc/backlog/c2b/services/MappingFiles.scala
@@ -7,8 +7,9 @@ import scala.collection.JavaConverters._
 import com.nulabinc.backlog.c2b.Config
 import com.nulabinc.backlog.c2b.core.Logger
 import com.nulabinc.backlog.c2b.datas.MappingContext
-import com.nulabinc.backlog.c2b.interpreters.{AppDSL, ConsoleDSL}
-import com.nulabinc.backlog.c2b.interpreters.AppDSL.AppProgram
+import com.nulabinc.backlog.c2b.dsl.AppDSL
+import com.nulabinc.backlog.c2b.dsl.AppDSL.AppProgram
+import com.nulabinc.backlog.c2b.interpreters.ConsoleDSL
 import com.nulabinc.backlog.c2b.persistence.dsl.{StorageDSL, StoreDSL}
 import com.nulabinc.backlog.c2b.serializers.{CSVRecordSerializer, Serializer}
 import com.nulabinc.backlog.c2b.services.DiffPatch.DiffMap
