+++ src/main/scala/com/nulabinc/backlog/c2b/syntax/BacklogResponseOps.scala
@@ -1,9 +1,10 @@
 package com.nulabinc.backlog.c2b.syntax
 
 import com.github.chaabaj.backlog4s.dsl.HttpADT.Response
+import com.nulabinc.backlog.c2b.dsl.AppDSL
+import com.nulabinc.backlog.c2b.dsl.AppDSL.AppProgram
 import com.nulabinc.backlog.c2b.exceptions.CybozuLiveImporterException
-import com.nulabinc.backlog.c2b.interpreters.AppDSL.AppProgram
-import com.nulabinc.backlog.c2b.interpreters.{AppDSL, ConsoleDSL}
+import com.nulabinc.backlog.c2b.interpreters.ConsoleDSL
 
 object BacklogResponseOps {
   implicit class ResponseOps[A](response: Response[A]) {
